renamed helpers to lib
