+++ lib/cardTypes.js
