+++ lib/clone.js
