+++ lib/findType.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var types = require('../cardTypes');
+var types = require('./cardTypes');
 
 function findType(type, customCards) {
   return customCards[type] || types[type];
