+++ lib/hasEnoughResultsToDetermineBestMatch.js
