+++ lib/isValidInputType.js
