+++ lib/loopOverCardPatterns.js
