+++ lib/matchesPattern.js
