+++ lib/matchesRange.js
